Fixing Missing Mapping things Pt.1 <!-- Write **BELOW** The Headers and **ABOVE** The comments else it may not be viewable. --> <!-- You can view Contributing.MD for a detailed description of the pull request process. -->
## About The Pull Request

<!-- Describe The Pull Request. Please be sure every change is documented or this can delay review and even discourage maintainers from merging your PR! -->

Changed an empty house that was never used into the prototype hideout/bar for the Daughters of Cacophony. This gives them a slightly wider buffer between themselves and the Toreador primogen.

Removed the brain numbing stripclub music for folks who actually enjoyed hanging out there, so they can either listen to or play their own.

## Why It's Good For The Game

<!-- Please add a short description of why you think these changes would benefit the game. If you can't justify it in words, it might not be worth adding. -->
These are things that I paid to be added ages ago, as well as removal of something that almost everyone who experienced it complained about.

## Changelog
:cl:
add: Added bar/hideout for DoC That I paid for over a year ago but was never added
add: Added keys to DoC spawns to be able to open the club for business, as well as to be able to access the dressing room.
sounddel: Removed the music from the stripclub that was doing permanent mental damage to all patron
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put your name to the right of the first :cl: if you want to overwrite your GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the icon ingame) and delete the unneeded ones. Despite some of the tags, changelogs should generally represent how a player might be affected by the changes rather than a summary of the PR's contents. -->