Increase the number of max save slots. (#384)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request

Increase the number of save slots from 10 to 20, 50 took too much of the
screen so i lowered.

20 Slots:


![image](https://github.com/user-attachments/assets/41e29ae7-c247-4325-80f8-b40324871089)

50 Slots:

![image](https://github.com/user-attachments/assets/b2d20aeb-0da1-431e-8b9d-7e68d571a92e)


## Why It's Good For The Game

The people want it , i also agree that 10 is too low.

## Changelog
:cl:
tweak: changed the number of save slots from 10 to 30.
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->