Clan clothes for sale (#440)
## About The Pull Request

Clan clothes for sale... simple as it is

## Why It's Good For The Game

Idk saw in the discord that it'd be cool and well... here it is
(wl clans weren't touched)
<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->

## Changelog


added almost all clan clothes to the shop
added balaclava to mask shop 
sorted clothes in shop lists