Examine additions (#458)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request
Adds varying examine texts if you look at a person - if they are of the
same faction as you, you will recognize them for their job. Otherwise,
there are varying levels of fame and such that allow you to be
recognized by your faction as a member of a specific clan, recognized by
only your clan as one, or famous enough that all kindred who meet you
know who you are, to represent that people are aware of who's in their
social circles.

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
It will make it easier for faction members to recognize each other's
social roles.
<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->

## Changelog
:cl:
add: Added new examine messages
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->

---------

Co-authored-by: TheCarnalest <167369312+TheCarnalest@users.noreply.github.com>