Adds greek Tragedy + greek Comedy masks (BY SPEFO) (#452)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request
Adds two masks to the mask shop, a Tragedy and a Comedy mask. All was
sprited and coded by Spefo; I just made the PR because their Github was
breaking.

![image](https://github.com/user-attachments/assets/fe99f0f3-2a79-4d03-a35b-44dfcb8279d8)

![image](https://github.com/user-attachments/assets/405caa87-a443-4442-8675-68a99c7870fc)

![image](https://github.com/user-attachments/assets/7a2b75c7-e32c-46e0-91a1-ba2e6b33eae6)

![image](https://github.com/user-attachments/assets/1050ebba-4e2a-4e77-8168-454c825d1c46)

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
Tragedy masks are part of Nosferatu's clan iconography, and also,
they're just kinda neat.
<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->

## Changelog
:cl:
add: Greek tragedy and comedy masks
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->