Tremere Anarchs & Ministry Anarchs  (#459)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request
For some reason, Ministry weren't allowed to join as Anarchs. This made
no sense, considering fucking gargoyles and kisayds can - as well as
many other exotic clans/bloodlines - when many members of the Followers
of Set already agree with the moral views of most Anarchs. Also Tremere
can now Be Anarchs but Not the Baron. I Encourage the Chantry to Hunt
Them Down and Kill Them and Shit.

It compiles. Works fine. It's literally just five lines of code.

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game Ministry Anarchs are cool. 
Tremere Antitribu Still Exist FUck You Fuck You Fuck You Fuck you Fuck
y-
<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->

## Changelog
:cl:
add: Gave Ministry the ability to play as Anarchs. 
add: Gave Tremere the ability to play as Anarchs
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->