Payday 2 DLC: Part 1 (#488)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request
Adds vault doors - currently they are lockable with electronic lock,
keys, or both. The combination lock as yet does not actually function.
The doors are unbreakable as you'd expect save for one small part -
thermal drills. Take this bad boy, put it down, fill it up with gas, and
turn it on. It will break the vault door over time - this is quite loud,
though, and is sure to invite attention.
Tested it for bugs in as many ways as I could.

![dreamseeker_JNyAkoavd8](https://github.com/user-attachments/assets/3ba8738c-54dc-4f3a-829c-09a2f860f000)


<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
It will expand upon the already existing bank that we've got and open a
path for future heisting updates.
<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->

## Changelog
:cl:
add: Added vault doors and thermal drills
soundadd: added multiple sounds for vault doors and thermal drills
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->