Damage overlay will now properly scale with max health. (#489)
You're playing a vampire with more than triple the health of an average character, and someone shoots your character! That's bad! What's worse, even though your character only took around 90 damage you can barely see shit because of the big red overlay on the screen, even though the damage taken is proportionally not that great.
This PR fixes it.