Require whitelists to learn certain Disciplines (#388)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request
This adds a check when teaching people Disciplines to see if they have
access to at least one Clan with the Discipline. This works to prevent
Disciplines from whitelisted Clans being shared with people who do not
have that whitelist.

Also does some code changes to make Clans and their code less weird,
like tweaking the Discipline lists. Also renames bruja.dm to brujah.dm
to correct the minor misspelling.
<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
People have been trying to learn Disciplines ingame as a way to bypass
whitelisting requirements. This is kind of unacceptable and something
had to be done about it.
<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->

## Changelog
:cl:
tweak: In order to be taught a Discipline, you now need access to at
least one Clan with that Discipline. This means whitelists are required
for certain Disciplines.
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->