Fixes Auspex not actually seeing in the dark (#462)
It couldn't actually see in the dark because the "see_in_dark" variable
was untouched, and it determines the distance after which tiles are
completely invisible and unobservable by the mob if they're dark enough.
This fixes it by adding actual behavior to TRAIT_NIGHT_VISION beyond
just adding a very thin layer of visibility for tiles that are revealed
in the dark while spawning eyes inside the player's head, so now you can
actually see what's in the sewers.

Might make Auspex brighter after this PR.