Fixes 460 spelling errors of "clan", git history preserved (#501)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request

Really just the title, though I did use git commands to properly
document the movement of the files to make sure the history of the repo
doesn't get fucked up. Please inquire within if you do anything moving
files around, I can show you how to do it right.

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game

i think it's good for us to spell good because we want people to think
we read good and do other stuff good too

<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->

## Changelog
:cl: Bisar
spellcheck: Amended file and code spelling errors in a git-friendly way
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->