Obfuscate hotfix; no more invisible murder [EMERGENCY HOTFIX, LIABLE FOR REVERT PENDING RETROACTIVE REVIEW] (#537)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request

Pending testing in ~2 hours from me PRing this: Obfuscate will end and
go into a cooldown of 60 seconds (down to a minimum of 30 seconds if you
have level 5+) if you perform an action with an object on anything.

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game

I was not aware that we had no code whatsoever to break Obfuscate if you
just went bananas with a gun and this needs to be addressed immediately.
I would merge now but I'm pushing this from work and still need to test
it.

<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->

## Changelog
:cl: Bisar
fix: Obfuscate is no longer an impenetrable invisibility cloak.
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->

---------

Co-authored-by: Metekillot <49173900+Metekillot@users.noreply.github.co>