Reduces Blood Heal's burn damage healing by 75%. (#437)
As a (likely better) alternative to #408, this PR makes the Blood Heal
ability heal 75% less burn damage instead of reducing all burn healing
for vampires by 80%, because healing a maximum of 80 burn damage (capped
at gen 9) is kind of ridiculous. It will now behave more like aggravated
damage, and this version will still allow vampires to go to the hospital
to get healed if they so wish.
Also cleans up the code a little.

---------

Co-authored-by: Joshua Kidder <49173900+Metekillot@users.noreply.github.com>