Expands what splats dealer can be (#574)
## About The Pull Request

After a conversation about a lack of roles for KJ, I gave it some
thought and looked at dealer and thought it would be interesting to let
Garou be a wise-guy glasswalker and run the warehouse, or let a black
spiral try to start a drug ring within the warehouse too. So this PR
effectively makes the warehouse a wild card and its own power through
economic might which seems to be the way of things moving into the
future.

## Why It's Good For The Game

Round diversity is, mostly, the number one goal here. The fact that the
warehouse isnt guaranteed to be Camarilla aligned will allow more
inter-round diplomacy and even the allowance of gimmicks on an easier
scale, like a glasswalker dealer of the wise-guy camp.

## Changelog

:cl:
Add: Allows werewolves & wk to be a Dealer.
Changes: The job text when you join.
/:cl:

---------

Co-authored-by: cheapcents <91615253+cheapcents@users.noreply.github.com>