Potence nerf and an unarmed combat change (#461)
Revival of #397 , this PR does two things:
1. It toggles the "gentle" variable to Potence's throw_impact, which
disables damage and stuns taken from colliding with other things such as
walls.
2. It reworks the punching knockdown system. Previously you needed
Potence to have a chance to knock someone down (until higher levels of
Potence would just be guaranteed to knock down and stunlock), now
punching has a 10% chance per punch to knock someone down for 2 seconds,
which increases up to 30% per punch depending on the difference between
the puncher's physique and whichever is greater of the target's physique
or dexterity. The difference for the stun chance increase has to be at
least 2 (example: user's physique of 5 vs target's physique and
dexterity of 3, which increases the stun chance to 20%).

I don't like how unatomic the PR has become.