Provides the Primogen with badges for Primogen Staff (#651)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

In line with the Counsel of Primogen sourcebook for VtM, I have provided
the Primogen with three badges in their backpack starting kit.

**Primogen's Myrmidon Badge**

A Myrmidon is the personal bodyguard of a Primogen, responsible for
their safety and security, as well as breaking legs when needed.

**Primogen's Steward Badge**

A Steward is responsible for maintaining, stocking, and staffing the
Primogen's domain/haven, hosting guests, etcetera.

**Primogen's Whip Badge**

A Whip is a Primogen's equivalent of a Seneschal, handling all of the
personal affairs of the Primogen delegated to them, and often
representing the Primogen at lesser functions.

## Why It's Good For The Game

The Primogen presently lack any structure to support maintaining a
'team' of confidants/staffers, which is extremely common for them to do
in the lore. Despite this, Primogen players tend to be extremely active
in game despite lacking the infrastructure individuals like the Prince /
Capo / Regent have to support them via underlings.

These three badges will enable and encourage Primogen to hire/recruit
other Kindred throughout the city to serve as their staff.

## Changelog

:cl:
add: Gave the Primogen three badges to distribute to their selected
servants in line with official lore book.
:cl: