Renames Scourge to Hound, adds Sheriff to their known contact (#671)
## About The Pull Request

- Renames Scourges to Hounds, replaces the vague "agent" name with
"hound" in code.
- Adds Sheriff to their known contacts.
- Changes their supervisor from Prince to Sheriff. (Their Department
Head remains the Prince.)

### Testing:

Job selection


![image](https://github.com/user-attachments/assets/65de9c95-0f52-491e-b91c-60f446871ab7)

Spawning in


![image](https://github.com/user-attachments/assets/0dc3e607-0fd3-40be-8047-eae03790a8da)

About me


![image](https://github.com/user-attachments/assets/af01b64e-43eb-4f23-a71b-1e329dea2fb4)

## Why It's Good For The Game

Scourges are appointed by Princes to hunt Thin-Bloods, Caitiffs, and
illegal Embraces, and they only answer to the Prince. Hounds are the
subordinates of the Sheriff.

Technically, we have the latter. The current job code is a bit
schizophrenic, at one point it says their supervisor is the Sheriff, but
then it appoints the Prince as their boss and only gives them their
contact.

(On another note, Hounds can both act as executioners and detectives,
while Scourges are murderhobos - it is better to reinforce their former
role. Players are already treating this role as Hounds.)

## Changelog

:cl:
fix: Renamed the Scourge role to Hound.
add: Hounds now answer to the Sheriff, instead of the Prince directly.
add: Hounds now also get the Sheriff's phone contact by default.
/:cl: