911 Dispatcher job for humans so we don't have to call personal phones in an emergency! (#423)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request

So this is a quick PR I've thrown together to add a new role for humans
(and I suppose some ghouls can slip through the net). A dispatcher!


![image](https://github.com/user-attachments/assets/cb2c8695-2d2d-4e92-ad07-3cbc9a53ddf0)

There are a couple of things to document on this one, first is the code
changes that mean that when 911 is dialed what is discussed (assuming a
dispatcher picks up) is broadcast to all the police radios and clinic
radios. The role has two slots, this should mean there is cover if one
of the dispatchers wants to stretch their legs for a short break/ meets
an unfortunate end.

Obviously I've had to create the dispatch office too... which I've
chosen the building attached to the PD for, it was the easiest choice
but I don't mind iterating on this. Inside the office is everything the
dispatcher needs to contact the police, they can link their radio via
the portal there.


![image](https://github.com/user-attachments/assets/9958baea-c28f-42ae-b951-b5d2fb091e39)

I've updated it so it also works over the new radio system.

**Not part of this PR but:** Have we considered converting the map type
away from the byond default? Will help with merge conflicts!

## Why It's Good For The Game

Well I can only speak personally here but I think that having this role
will help the human roles feel a tad more rounded, I've had a few
moments of dissonance where I've wanted to call emergency services as a
human and had basically no number I can call because the officers are
all super busy getting murdered and the clinic staff are all vampires so
they don't want to know.

Also I think it'd be a nice fun role too, not as stressful as playing
police as you're hardly an active combatant but still involved with
what's going on!

Couple more Screenshots:


![image](https://github.com/user-attachments/assets/d65dd875-ca85-474b-b87b-5f67abd9e47e)

![image](https://github.com/user-attachments/assets/d69f07ff-2173-402a-a568-023fb8ef44b0)



- If people are interested in more like this I might add a news crew and
news reporter too!

## Changelog
:cl:
add: Added the dispatcher role
add: Added the dispatch phone that is directly tied to the police radio
as well
tweak: the building attached to the PD is now a dispatchers office.
/:cl:

---------

Co-authored-by: TheCarnalest <167369312+TheCarnalest@users.noreply.github.com>