Ventrue Domain Remap (#695)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request

This PR remaps the whole of the Ventrue Domain that is within Millennium
Tower. The goal was to make the Domain more attractive to roleplay
within, especially amongst Ventrue themselves who are supposed to be
highly connected and group-minded. The colour theme has been changed
from the hideous green it was to royal purple, fitting for the Clan of
Kings.

**Primogen's Office**

The Primogen, with the cover title of 'Board Chair' on his doors, has a
new office separate from his bedroom. The office has a view of the city
and a view of the board room directly outside. It bears two medals, one
an olympic medal and one a world war two medal. There are two busts, one
of Hippocrates and one re-named and re-described of Antonius of Gaul (a
Ventrue Methuselah!)

The office features a fancy pen, a secure briefcase, and a button to
control his curtains for private meetings.

![image](https://github.com/user-attachments/assets/a453f8c2-2d8a-453a-9861-502d3a66a8d4)

**Primogen's Bedroom**

The Primogen's Bedroom has been modified to be more open. It has a
curtained gun cabinet containing the Primogen's special rifle and its
ammunition, an elevated coffin bed, and a cabinet with some various
suits/fancy clothes, as well as purple robes and a fancy venetian mask.

![image](https://github.com/user-attachments/assets/2c4e248d-52a4-48fc-8499-634b2a02e824)

**The Gerousia Meeting Room aka the Board Room**

This is the room for the Primogen to hold meetings of Ventrue, themed
towards being the Ventrue's city-based organizational unit of a
'Gerousia' often called a 'Board' in modern nights. The room is meant to
be grandiose and minimalist, with all of the focus on the large fancy
table in the middle and the important people sitting at it. The extra
space leaves room for attendants/servants/visitors to stand at the side
to listen.

![image](https://github.com/user-attachments/assets/ecf1e856-53d8-4dae-b33b-f397d125d0d4)

**Reception**

The entry-way has been expanded significantly into a reception area. It
now features a desk for the primogen's steward or a particularly servile
Ventrue to sit at and receive guests.

![image](https://github.com/user-attachments/assets/f4cc472e-7e5f-435a-81cf-d7bca69414ab)

**Ventrue Offices**

The bedrooms from the previous iteration have been replaced with offices
featuring desks, computers, some fancy but less fancy than the
Primogen's fountain pens, and cabinets of some suits and skirts and
purple robes w/ regular venetian masks. There are three of these offices
available along the hallway.

![image](https://github.com/user-attachments/assets/2f40cde6-7d12-4112-a973-2df7f208a572)

**Staff Wing**

The corridor connecting reception, the board room, and the ventrue
offices together. It features some sitting areas, a single elite blood
bag, and decorative things.

![image](https://github.com/user-attachments/assets/ad1b522f-41f5-4522-a15a-f5e6916b0044)

**Storage Closet**

A fluff storage closet with nothing particularly useful has been added.
It contains a plastic lawn chair 'of shame' for the Ventrue Primogen to
embarrass particularly ignoble Ventrue with by requiring they sit in it
at meetings. It also has a lil cockroach.

![image](https://github.com/user-attachments/assets/0b37d9a2-d01c-404d-86dc-06bc09ab0ee5)

## Why It's Good For The Game

The Ventrue Domain is infamous amongst Ventrue-mains for being god
awful. It's tight, small, ugly, and has little to no utility. Based upon
feedback received during the mapping process, this refit is much desired
and will make the domain a more enjoyable place to roleplay in (or break
into!). It also makes the Ventrue seem as important and organized as
they are supposed to be.

<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->

## Changelog

<!-- If your PR modifies aspects of the game that can be concretely
observed by players or admins you should add a changelog. If your change
does NOT meet this description, remove this section. Be sure to properly
mark your PRs to prevent unnecessary GBP loss. You can read up on GBP
and its effects on PRs in the tgstation guides for contributors. Please
note that maintainers freely reserve the right to remove and add tags
should they deem it appropriate. You can attempt to finagle the system
all you want, but it's best to shoot for clear communication right off
the bat. -->

:cl:
map: Remapped the Ventrue Primogen's Office into a proper Ventrue
Domain.
add: Created a new type of button-controlled curtain.
:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->

---------

Co-authored-by: Miraviel <33333517+Miraviel@users.noreply.github.com>