Dispatcher radio follow up (#709)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request

Just a quick follow-on PR for the new dispatcher job. I've made it so
the dispatcher radios the police now, rather than them hearing the call.

Also fixed the NPC witness system spamming radio chatter for every
radio.


![image](https://github.com/user-attachments/assets/daf33eac-5ac0-4406-9e96-ccadc2bfd5b9)

![image](https://github.com/user-attachments/assets/0d361607-8782-4a08-ab4c-123901f40def)


## Why It's Good For The Game

Less radio spam!

<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->

## Changelog

<!-- If your PR modifies aspects of the game that can be concretely
observed by players or admins you should add a changelog. If your change
does NOT meet this description, remove this section. Be sure to properly
mark your PRs to prevent unnecessary GBP loss. You can read up on GBP
and its effects on PRs in the tgstation guides for contributors. Please
note that maintainers freely reserve the right to remove and add tags
should they deem it appropriate. You can attempt to finagle the system
all you want, but it's best to shoot for clear communication right off
the bat. -->

:cl:

fix: fixed the dispatch radio chatter

/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->

---------

Co-authored-by: TheCarnalest <167369312+TheCarnalest@users.noreply.github.com>