Bear - The Return of the Bearening (#719)
## About The Pull Request

Mistakes were made and I accidentally removed the PR. It's fine. 

We've also removed the actual bear cave, mainly because of the grief
potential and simple mob behaviour. For now, it's an Admin Spawn.

## Why It's Good For The Game

Admin tomfoolery. Maybe in the future when Simple Mobs are updated we'll
return the Bear to it's natural habitat, but until then it's just
something that the Admins can throw at a random shitter.

## Changelog


:cl:
add: Bear
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->

---------

Co-authored-by: Adri <33333517+Miraviel@users.noreply.github.com>