Removes stray diff file and ports VtM icon in tgui backgrounds (#959)
I may or may not have forgotten to remove the diff file that I used to
port the 516 changes from TFN. in #956. Anyway whoever left it there, I
removed it :P

Also ports https://github.com/Near-Web/The-Final-Nights/pull/216 which
is a minor change that makes the intrfaces look way better.