Home Invasions and individual civilian dwelling areas (#954)
This PR makes a separate area for the 31 civilian dwellings found on the
map, adds a set of crates, windows, doors and alarm boxes to each one of
them (and removes the civilian ghost spawns that were there. Should this
functionality be wanted back, there are other ways of achieving that,
but it seems irrelevant considering we allow Respawns) and ties it to a
home invasion/home alarm mechanic.

This also adds some other minor mechanics like curtains that don’t
require pixel_x/y manipulation to fit on one side of the door and
creates uniquely named areas for each of the 31 dwellings.

This system is relatively self-contained and relies on introducing a few
dummy processes to existing commands and lives on its own set of
definitions to make porting/reattaching easier (hopefully anyway)

Loot “category” for each house and spreads of loot within is randomized
at roundstart, according to a few gvars that likely will ultimately be
just a proportion. This lets admins make last minute edits/additions to
the system in the future.

Looking at an entryway of a civilian dwelling with a lockpick in one’s
hand begins the casing process, which upon completion gives information
about the alarm system and loot quality inside.

Loud actions like breaking dwelling windows or kicking the door down
trigger the alarm loop automatically, otherwise it slowly fills up while
the player(s) looks through the containers found in each house.

Looting is done by finding and using special containers around the
house. Each has a set number of “tries” and set number of “successes”
assigned to them, a success means that loot is rolled from the currently
active table and dispense to the player. The time required to complete
the search is directly proportional to the amounts of “tries” left – the
more a container is searched, the easier it gets to search it more.
Reaching a 0 in successes terminates the action.

Once enough ‘heat’ is accumulated from actions inside the house, the
alarm system activates and beings a security scan. At this point,
carefully watching the alarm system for a moment where one of its diodes
does not light up and using a lockpick on it either disables the system
outright or resets the scan (ideally at this point one player will
dedicate their time to maintaining the security system)

Failing to click at the right time or running out of time makes the
alarm report the intrusion to the PD. From now on, every action that
generates heat also sends another alarm. This can be overridden with a
special keycard available to cops (one in the Chiefs locker, one in the
PD Armory)