Delivery Minigame (#969)
This PR introduces a truck driving/delivery minigame and some associated
mechanics that both intend to provide opportunities for RP together (the
missions can be undertaken by multiple people, in fact they are
suggested for the longer variants, also the more people, the faster the
turnaround); some indirect competition, for the time moment only really
presented via end-of-round statistics comparing delivery to the three
outposts available.

Once taken, a contract spawns a special truck and can be viewed to look
for dispensers that distribute cargo of a specific type. Once meeting
the quotas at all receiver, the truck should be taken back to the garage
of origin and the contract finalized. Payment is distributed to bank
accounts of participants.

Extra helpers (that get keys that allow them to actually dispense and
deliver crates) can be added by hitting them in the head with the
contract item. They can also be removed the same way, though I'll likely
block that once the contract is finalized to prevent shenanigans.